Add manifest schema.
West manifests are used to specify a project configuration, i.e., what
Git repositories to check out to install Zephyr, where to get them
from, what revisions to check out, and where to check them out
locally.

Manifests are written in YAML. This commit adds a pykwalify schema
which can be used to do basic validation of a manifest's structure.
It also provides documentation in comments on the keys in a manifest
and their meaning.

Signed-off-by: Marti Bolivar <marti@opensourcefoundries.com>